Changed NuGet package for MySql connection
