+++ Infrastructure/Configuration/Provider/IDbAccessProvider.cs
@@ -1,6 +1,9 @@
-﻿namespace Infrastructure.Configuration.Provider;
+﻿using Microsoft.EntityFrameworkCore;
+
+namespace Infrastructure.Configuration.Provider;
 
 public interface IDbAccessProvider
 {
     string GetConnectionString();
+    ServerVersion GetServerVersion();
 }
