+++ Infrastructure/DbAccess/EfDbContext/DataContext.cs
@@ -17,6 +17,7 @@ public class DataContext : DbContext
     protected override void OnConfiguring(DbContextOptionsBuilder optionsBuilder)
     {
         var connectionString = _dbAccessProvider.GetConnectionString();
-        optionsBuilder.UseMySQL(connectionString);
+        var serverVersion = _dbAccessProvider.GetServerVersion();
+        optionsBuilder.UseMySql(connectionString, serverVersion);
     }
 }
