+++ WebDiplomaWork/.gitignore
@@ -27,6 +27,9 @@ Obj/
 # Visual Studio 2015 cache/options directory
 .vs/
 
+# Rider options directory
+.idea/
+
 # MSTest test Results
 [Tt]est[Rr]esult*/
 [Bb]uild[Ll]og.*
