+++ WebDiplomaWork/appsettings.json
@@ -9,10 +9,11 @@
   "GeneralConfiguration": {
     "DbConfiguration": {
       "Port": 3306,
-      "Host": "91.228.155.252",
-      "DatabaseUserName": "",
-      "DatabaseName": "",
-      "DatabasePassword": ""
+      "Server": "91.228.155.252",
+      "ServerVersion": "10.11.4",
+      "UserId": "",
+      "Database": "",
+      "Password": ""
     }
   },
   "AllowedHosts": "*"
